hack: Pass OTP to the server
The plan to fix this hack is to pass a LoginKeys parameter to the server
functions. This parameter will be a struct that contains
authId, passwordAuth, and otp. This will save the Login system from
needing to touch gLobbyCache.

Once we have non-blocking URL queries, we can pass the Lobby or Login
directly. We can't do that now, since that would lock the login mutex.