Modernize the Login object & store the authKey
Before, the authKey would be loaded from disk on each server request.
Now, it lives inside the Login object. To make this possible,
the Login object now follows the same structure as the Wallet object.

This also fixes a bug which would cause account activation to fail
due to a missing authKey.