Do not calculate balance in ABC_GetWalletInfo
Balance info is a heavy calculation in the core as it has to iterate over all transactions. The app will have to call ABC_GetTransactions anyway, so we can just use the balance tally after that's called. Ideally, ABC_GetTransactions should just return a balance so the GUI isn't calculating it.