Simplify libbitcoin-server broadcasts
The previous implementation had race conditions,
where transactions wouldn't show up in the database until some time later.
The GUI would ask to see the recently-sent transaction,
and it would come back empty.

This separates adding the transaction to the database from the broadcast.
That way, the GUI doesn't have to wait for the broadcast to finish before
the transaction is available.