Re-write the transaction cache update logic
The goal is to make "clearing the cache" a thing of the past.
The cache is now explicitly aware of which transactions are missing,
and uses this to-do list to drive the updater process.

The address cache is responsible for storing lists of txids for each address.
To make this possible, the on-disk storage format has been changed to JSON.
This will make future upgrades much easier, and also helps with debugging.

Saving the address cache to disk also makes it possible to preserve the
last-check timestamps. Now, when the app restarts, it can actually
pick up where it left off, instead of starting the checks from scratch.
This will should a major impact on large wallets on slow connections.

Both Stratum and libbitcoin include heights in their address histories,
the updater no longer fetches heights as a separate step.
This cuts down on the number of queries, but also means that not all txs
have known heights. To handle this, the tx cache now stores heights
in a separate table from the raw transaction data.

With this enhanced knowledge of what is done and what remains to be done,
the cache is now much more picky about what it tells the GUI.
It will only report tranactions that are fully downloaded,
including their inputs. This means that balances will always be correct,
at least on a per-transaction basis.

The address cache now provides an address-done callback.
This means that sweeps can finally be fast and accurate,
since the wallet now knows the exact moment when the address is finished.