Do not auto-create JSON array and object pointers
These types now initialize themselves in the constructor,
then remain pointing to the same JSON value through their lifetimes.
Besides being simpler, this also makes the classes more predictable.
A `JsonObject` that is constructed from a `JsonPtr` will always
have the same contents, and won't mysteriously re-create itself.

This might be a tiny bit slower in the case of a load,
since the initial JSON value will be replaced with the loaded value.
This can be avoided by passing `nullptr` to the constructor.