Implement the new Airbitz fee algorithm
Before, the wallet would send fees right away.
Now, the wallet can allow fees to accumulate for a time before sending them.
This is especially important for incoming fees,
where sending a second transaction for every receive would be impractical.

Each transaction now carries an `airbitzFeeWanted` field,
which holds the amount owed. This can be different from the `airbitzFeeSent`
field, which holds the amount actually paid.

The wallet can take the difference between the total fees owed and
and the total fees paid to see how many fees are remaining.
If this number is large enough for a long enough time,
the wallet will react by sending a fee transaction.