Move non-user data out of `TxMetadata`
This structure has had a long and messy history,
but its original design goal was to contain user-settable metadata.
Since then, various transaction and request related things have snuck in.
This commit moves those things back out to the address and transaction
storage layers.

The data saved to disk should be the same as before,
as well as the public C API. This is purely an internal clean-up.