Fix the max-spend calculation
The previous algorithm assumed that we wanted to sweep the wallet,
but this can be uneconomical in wallets with a lot of dust.
Instead, just try a bunch of different values in a binary search.
The search will always find a local maximum,
but maybe not the global maximum if the fees aren't monotonic.