Munge the block time into `TxInfo`
If the block confirmation time is earlier than the metadata timestamp,
use that instead. We must have seen the transaction after it was confirmed.

If the transaction is unconfirmed, or if the block header isn't available,
we use the current time as the fallback block timestamp. In this case,
the metadata timestamp will almost always win (assuming we have metadata).