deps: Update timestamps when extracting tar's
Doing a fresh unpack should touch the timestamps on all files.
Otherwise, the build tools won't see that the sources have changed,
and might leave local changes sitting around.

This bit us when trying to remove a patch from the libgit2 recipe.
Extracting the archive created a timestamp in the past,
so the build step didn't recompile the file,
even though it no longer contained the patch.