Provide a way to log in with a decryption key
In the Airbitz app, "logging in" is defined as obtaining the decryption key.
If the key can be stored securely outside of Airbitz the app,
such as in a fingerprint reader, we don't need anything else to log in.

This login method is offline-only for now,
at least until the server supports some kind of pubkey-based authenticaiton.